Update readme (#8)
* Update readme file. Move blendshapes enum into different file
