+++ Assets/Rokoko/Scripts/Core/ARKitBlendshapes.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9f43a84310e40244c970e95e1d934b1a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
