+++ Assets/Rokoko/Scripts/Editor/BlendShapesMappingEditor.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using UnityEditor;
 using UnityEngine;
 using Rokoko.Helper;
+using Rokoko.Core;
 
 namespace Rokoko.UnityEditor
 {
