+++ Assets/Rokoko/Scripts/Mono/RokokoHelper.cs
@@ -6,63 +6,6 @@ using UnityEngine;
 
 namespace Rokoko.Helper
 {
-    [System.Serializable]
-    public enum BlendShapes
-    {
-        eyeBlinkLeft = 0,
-        eyeLookDownLeft = 1,
-        eyeLookInLeft = 2,
-        eyeLookOutLeft = 3,
-        eyeLookUpLeft = 4,
-        eyeSquintLeft = 5,
-        eyeWideLeft = 6,
-        eyeBlinkRight = 7,
-        eyeLookDownRight = 8,
-        eyeLookInRight = 9,
-        eyeLookOutRight = 10,
-        eyeLookUpRight = 11,
-        eyeSquintRight = 12,
-        eyeWideRight = 13,
-        jawForward = 14,
-        jawLeft = 15,
-        jawRight = 16,
-        jawOpen = 17,
-        mouthClose = 18,
-        mouthFunnel = 19,
-        mouthPucker = 20,
-        mouthLeft = 21,
-        mouthRight = 22,
-        mouthSmileLeft = 23,
-        mouthSmileRight = 24,
-        mouthFrownLeft = 25,
-        mouthFrownRight = 26,
-        mouthDimpleLeft = 27,
-        mouthDimpleRight = 28,
-        mouthStretchLeft = 29,
-        mouthStretchRight = 30,
-        mouthRollLower = 31,
-        mouthRollUpper = 32,
-        mouthShrugLower = 33,
-        mouthShrugUpper = 34,
-        mouthPressLeft = 35,
-        mouthPressRight = 36,
-        mouthLowerDownLeft = 37,
-        mouthLowerDownRight = 38,
-        mouthUpperUpLeft = 39,
-        mouthUpperUpRight = 40,
-        browDownLeft = 41,
-        browDownRight = 42,
-        browInnerUp = 43,
-        browOuterUpLeft = 44,
-        browOuterUpRight = 45,
-        cheekPuff = 46,
-        cheekSquintLeft = 47,
-        cheekSquintRight = 48,
-        noseSneerLeft = 49,
-        noseSneerRight = 50,
-        tongueOut = 51,
-        size = 52
-    }
 
     public static class RokokoHelper
     {
